+++ packages/runtime/adapters/replicators/replicator-via-api-handler/src/resume.ts
@@ -1,8 +1,13 @@
 import type { ExternalMethods } from './types'
 
-const resume: ExternalMethods['resume'] = async (pool, readModelName, next) => {
+const resume: ExternalMethods['resume'] = async (pool, readModelName) => {
   await pool.setReplicationPaused(pool, false)
-  await next()
+  return {
+    type: 'build-direct-invoke',
+    payload: {
+      continue: true,
+    },
+  }
 }
 
 export default resume