+++ packages/runtime/runtimes/runtime-single-process/src/compile-time/index.ts
@@ -1,0 +1,6 @@
+type ResolveConfig = any
+
+export const adjustResolveConfig = async (resolveConfig: ResolveConfig) => {
+  resolveConfig.target = 'local'
+  resolveConfig.externalDependencies = []
+}