+++ packages/runtime/runtimes/runtime-single-process/test/wrap-api-handler.unit.test.ts
@@ -214,7 +214,10 @@ describe('API handler wrapper for express.js', () => {
   }
 
   it('should work with primitive JSON handler with GET client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiJsonHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiJsonHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -284,7 +287,10 @@ Object {
   })
 
   it('should work with primitive JSON handler with POST client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiJsonHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiJsonHandler as any,
+      getCustomParams
+    )
     resolveHttpBody([
       Buffer.from('Body partition one'),
       Buffer.from('Body partition two'),
@@ -357,7 +363,10 @@ Object {
   })
 
   it('should work with text handler with any client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiTextHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiTextHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -427,7 +436,10 @@ Object {
   })
 
   it('should work with custom handler with any client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiCustomHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiCustomHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -497,7 +509,10 @@ Object {
   })
 
   it('should work with file handler with any client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiFileHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiFileHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -567,7 +582,10 @@ Object {
   })
 
   it('should work with redirect handler with any client request', async () => {
-    const wrappedHandler = wrapApiHandler(apiRedirectHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiRedirectHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -693,7 +711,10 @@ Object {
   })
 
   it('should work with empty end', async () => {
-    const wrappedHandler = wrapApiHandler(apiEmptyEndHandler, getCustomParams)
+    const wrappedHandler = wrapApiHandler(
+      apiEmptyEndHandler as any,
+      getCustomParams
+    )
     resolveHttpBody(null)
     await wrappedHandler(expressReq, expressRes)
 
@@ -764,7 +785,7 @@ Object {
 
   it('should work with empty end using chaining', async () => {
     const wrappedHandler = wrapApiHandler(
-      apiEmptyEndChainingHandler,
+      apiEmptyEndChainingHandler as any,
       getCustomParams
     )
     resolveHttpBody(null)